LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-named-scroll-progress-timeline.tentative-expected.txt


PASS scroll-timeline-name is referenceable in animation-timeline on the declaring element itself
FAIL scroll-timeline-name is referenceable in animation-timeline on that element's descendants assert_equals: expected "100px" but got "none"
FAIL scroll-timeline-name is not referenceable in animation-timeline on that element's siblings assert_equals: Animation with unknown timeline name holds current time at zero expected "50px" but got "none"
PASS scroll-timeline-name on an element which is not a scroll-container
FAIL Change in scroll-timeline-name to match animation timeline updates animation. assert_equals: expected null but got object "[object DocumentTimeline]"
FAIL Change in scroll-timeline-name to no longer match animation timeline updates animation. assert_true: Failed to create animation expected true got false
FAIL Timeline lookup updates candidate when closer match available. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'anim.timeline')"
FAIL Timeline lookup updates candidate when match becomes available. assert_equals: expected "50px" but got "none"
FAIL scroll-timeline-axis is block assert_equals: expected "50" but got "auto"
FAIL scroll-timeline-axis is inline assert_equals: expected "50" but got "auto"
FAIL scroll-timeline-axis is x assert_equals: expected "50" but got "auto"
FAIL scroll-timeline-axis is y assert_equals: expected "50" but got "auto"
FAIL scroll-timeline-axis is mutated assert_equals: expected "25" but got "auto"
